[move] 프로필 엔티티 패키지 안으로 이동 #76
