+++ wingle/src/main/java/kr/co/wingle/profile/ProfileRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import kr.co.wingle.member.entity.Member;
+import kr.co.wingle.profile.entity.Profile;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
 	@Query("select p.nation from Profile p where p.member = :member")
